Update Lesson_4_Checking_For_Character_NFT.md
On line 168 the variable holding the results from the await function is called `txn`. 

As you go through and talk about the code, the name of the variable has now changed to `characterNFT`. The discussion about the code needs to match the actual code.

Same problem on line 206 and 208 in that the variable txn has been replaced with characterNFT.