Update Lesson_3_Let_Users_Mint_Your_NFT.md
This proposed change to the content is to reflect updated code under "Build a Mint NFT button section. The updated code fixes a bug that was allowing users to pass to the DAO member page even if they did not complete the NFT minting process. This change is a slight revision from a PR request by another user [mintNFT Bugfix #2] - that PR request has an errant ";" before ".catch((err) that caused a syntax error in replit and it wouldn't compile. this change removes the problem ";"