adding keccak256 -- pseudo-random number gen
I added a quick bonus section on incorporating keccak256 hashing to spice up the attack functions, as well as a call to action to refactor the codebase to provide for dynamic 'luck' or 'defense' character attributes. I feel it would be an excellent refresher for most people at this point in the course to go back, and make the changes needed for adding a luck or defense stat!